serialize outgoing peer requests on startup
We noticed on a testnet with 1 or 2 other available peers the node would
make many (up to 8) connection requests to the same peer. This resulted
in our node getting banned by the other peers because of the spamming.

Simply serializing the call for NewConnReq significantly reduces duplicate
connections. Sometimes we may see a duplicate connection on startup,
probably due to some race condition in the call cm.cfg.GetNewAddress.

This fixes the startup related issue https://github.com/gcash/bchd/issues/448,
and may also fix PR https://github.com/gcash/bchd/pull/296 but not sure the
entire context for that PR.