prevent sm.msgChan from blocking server.peerHandler
Sometimes SyncManager.blockHandler goroutine gets backed up, probably
caused by large blocks or some other unknown reason. So this update ensures
server.peerHandler does not get blocked when this happens.  Fixing this
blocking issue on peerHandler will allow Stop() to be called on connManager
syncManager, and addrManager on a shutdown request.

This seems to be the only place where peerHandler goroutine can be
possibly blocked by the state of the sm.blockHandler.